% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_columns.R
\name{arrange_cols}
\alias{arrange_cols}
\title{Move a set of columns to the left-hand side of a data frame}
\usage{
arrange_cols(df = df, refcols = refcols)
}
\arguments{
\item{df}{A data frame}

\item{refcols}{A vector of the column names you would like to move to the left side of the dataframe (in order)}
}
\description{
Move a set of columns to the left-hand side of a data frame
}

