% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocman.R
\name{bioc_install}
\alias{bioc_install}
\title{Bioconductor package management}
\usage{
bioc_install(lib.loc = .libPaths()[1])
}
\arguments{
\item{.libwork}{work R Library location}
}
\description{
Bioconductor package management
}
\keyword{Bioconductor}
\keyword{install}
\keyword{installation}
\keyword{package}

