% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_help.R
\name{build_help}
\alias{build_help}
\title{Build Static HTML Help with knitr}
\usage{
build_help(pkgname)
}
\arguments{
\item{pkgname}{package for which you want to build static html help}
}
\description{
Build Static HTML Help with knitr
}

