% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrtable.R
\name{corrtable}
\alias{corrtable}
\title{fileTable}
\usage{
corrtable(df)
}
\arguments{
\item{df}{dataframe suitable for conversion to correlation matrix}
}
\description{
fileTable
}
\examples{
df <- mtcars[, 1:7]
corrtable(df)
}
\keyword{(from}
\keyword{a}
\keyword{create}
\keyword{dates}
\keyword{file}
\keyword{filename)}
\keyword{files}
\keyword{folder}
\keyword{in}
\keyword{of}
\keyword{table}
\keyword{with}

