% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran}
\alias{cran}
\title{Install from CRAN}
\usage{
cran(x, .libwork)
}
\arguments{
\item{x}{package name in quotes}

\item{.libwork}{work R Library location}
}
\description{
Install from CRAN
}
\keyword{installation}
\keyword{package}

