% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cal.R
\name{create_cal}
\alias{create_cal}
\title{create_cal}
\usage{
create_cal(start_yr, end_yr)
}
\arguments{
\item{start_yr}{4-digit integer start year (no quotes needed)}

\item{end_yr}{4-digit integer end year (no quotes needed)}
}
\description{
create_cal
}
\examples{
cal_test <- create_cal(start_yr = 2005, end_yr = 2016)
View(cal_test)

}
\keyword{and}
\keyword{create}
\keyword{dates}
\keyword{df}
\keyword{end}
\keyword{monthly}
\keyword{of}
\keyword{start}

