% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tbl_df.R
\name{create_tbl_df}
\alias{create_tbl_df}
\title{Create a Basic Table DF (tbl_df)}
\usage{
create_tbl_df(fill_value, nrows, ncols)
}
\arguments{
\item{fill_value}{Default value for all data frame 'cells'}

\item{nrows}{Number of rows desired}

\item{ncols}{Number of columns desired}
}
\description{
Create a Basic Table DF (tbl_df)
}
\examples{
create_tbl_df(0, 1, 5)
}
\keyword{dataframe}
\keyword{make}
\keyword{results}

