% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict.R
\name{data_dict}
\alias{data_dict}
\title{build an initial data dictionary for a dataframe}
\usage{
data_dict(df)
}
\arguments{
\item{df}{dataframe whose columns you would like to list and class}
}
\description{
build an initial data dictionary for a dataframe
}
\examples{
df <- mtcars
dd <- data_dict(df)
dd
}

