% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTable.R
\name{fileTable}
\alias{fileTable}
\title{fileTable}
\usage{
fileTable(fpath = dd, search.pattern = "")
}
\arguments{
\item{search.pattern}{a string indicating search qualifier}

\item{folderpath}{the directory to search, defaults to dd}
}
\description{
fileTable
}
\examples{
ft <- fileTable("~")
ft
}
\keyword{(from}
\keyword{a}
\keyword{create}
\keyword{dates}
\keyword{file}
\keyword{filename)}
\keyword{files}
\keyword{folder}
\keyword{in}
\keyword{of}
\keyword{table}
\keyword{with}

