% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtext.R
\name{findtext}
\alias{findtext}
\title{findtext}
\usage{
findtext(vec_name, string, sel = "max")
}
\arguments{
\item{vec_name}{vector to search within}

\item{string}{text to locate, in quotes}

\item{sel}{supports either 'min' or 'max}
}
\description{
findtext
}
\keyword{a}
\keyword{character}
\keyword{find}
\keyword{max}
\keyword{min}
\keyword{of}
\keyword{or}
\keyword{position}
\keyword{string}
\keyword{text}
\keyword{vector}
\keyword{within}

