% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixlabels.R
\name{fixlabels}
\alias{fixlabels}
\title{fixlabels:  Standardize a vector of character values for use as dataframe column names}
\usage{
fixlabels(x)
}
\arguments{
\item{x}{character vector whose elements you want to standardize}
}
\description{
fixlabels:  Standardize a vector of character values for use as dataframe column names
}
\examples{
varnames <-  dff(variable = names(iris))
varnames
varnames <- varnames \%>\% 
  mutate(variable = fixlabels(variable))
varnames
}

