% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixnames.R
\name{fixnames}
\alias{fixnames}
\title{standardize column names for a dataframe}
\usage{
fixnames(df)
}
\arguments{
\item{df}{dataframe whose colnames you want to standardize}
}
\description{
standardize column names for a dataframe
}
\examples{
# Handling of periods in names:
# Assume that a period should be replaced by an underscore
# to preserve whatever separation of words was intended
# by the original inclusion of periods,  Don't
# want to simply remove them, but they're inconvenient.
# Replace periods with an underscore:
df <- mtcars
head(df)
names(df) <- paste0("_.", names(df), ")")
head(df)
names(df) <- fixnames(df)
head(df)
df
}

