% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_name_and_save_path.R
\name{listname}
\alias{listname}
\title{Format list name and save path}
\usage{
listname(list_name, time = time_stamp())
}
\arguments{
\item{list_name}{Desired list name, no trailing spaces required}

\item{full_listname}{List name with extension (as returned by listname function)}
}
\description{
Format list name and save path
}

