% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natable.R
\name{natable}
\alias{natable}
\title{Summarize Number of NAs by Variable}
\usage{
natable(df)
}
\arguments{
\item{df}{data frame from which you want the NA table}
}
\description{
Summarize Number of NAs by Variable
}
\examples{
natable(mtcars)
}

