% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclus.R
\name{nclus}
\alias{nclus}
\title{Evaluate optimal k for clustering}
\usage{
nclus(data, min_k = 2, max_k = 10)
}
\arguments{
\item{data}{data frame of scaled variables in columns, obs in rows, no row labels}

\item{min_k}{minimum k to allow}

\item{max_k}{maximum k to allow}
}
\description{
Evaluate optimal k for clustering
}
\keyword{by}
\keyword{criteria}
\keyword{k}
\keyword{selection}

