% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{numeric_summary}
\usage{
numeric_summary(df)
}
\arguments{
\item{df}{a dataframe containing at least one numeric column}
}
\description{
numeric_summary
}
\examples{
df <- numeric_summary(mtcars)
head(df)
}
\keyword{a}
\keyword{columns}
\keyword{create}
\keyword{dataframe}
\keyword{numeric}
\keyword{of}
\keyword{summarizing}
\keyword{table}
\keyword{the}

