% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctrankmat.R
\name{pctrankmat}
\alias{pctrankmat}
\title{Create a percent-ranked data frame from 
a subset of columns of an existing data frame}
\usage{
pctrankmat(df, begin, numdigits)
}
\arguments{
\item{df}{A data frame with one or more numeric columns}

\item{begin}{The first numeric column.  All columns beyond this one must be numeric.}
}
\description{
Create a percent-ranked data frame from 
a subset of columns of an existing data frame
}
\keyword{The}
\keyword{digits}
\keyword{for}
\keyword{in}
\keyword{number}
\keyword{numdigits}
\keyword{of}
\keyword{results}
\keyword{rounding}
\keyword{significant}
\keyword{the}
\keyword{use}

