% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLinks.R
\name{processLinks}
\alias{processLinks}
\title{Remove Links from a String}
\usage{
processLinks(x)
}
\arguments{
\item{df}{a dataframe}

\item{column.name}{name of the column to process, in quotes}
}
\description{
Remove Links from a String
}

