% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projdirs.R
\name{projdirs}
\alias{projdirs}
\title{Create Default Folder Structure for a New Project}
\usage{
projdirs(projectname)
}
\arguments{
\item{projectname}{name of the new project}
}
\description{
Create Default Folder Structure for a New Project
}
\keyword{project}
\keyword{workflow}

