% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{rdf}
\alias{rdf}
\title{read feather file}
\usage{
rdf(dfname, directory = here::here("feather"))
}
\arguments{
\item{dfname}{name of the file, in quotes, without an extension}

\item{directory}{directory to read from}
}
\description{
read feather file
}
\examples{
library(smutilities)
df <- iris
fd <- here::here("feather")
wf(df, "iris", fd)
df <- rdf("iris", fd)
}

