% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonconv.R
\name{remove_nonconv}
\alias{remove_nonconv}
\title{Remove Nonconvertible Text}
\usage{
remove_nonconv(df, column.name)
}
\arguments{
\item{df}{a dataframe}

\item{column.name}{name of the column to process, in quotes}
}
\description{
Remove Nonconvertible Text
}

