% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNA.R
\name{setNA}
\alias{setNA}
\title{Manage NA Values}
\usage{
setNA(df, na_val)
}
\arguments{
\item{df}{the data frame you're working with}

\item{na_val}{what should NAs be changed to? (e.g. 0, "", NA)}
}
\description{
Manage NA Values
}
\keyword{project}
\keyword{workflow}

