% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfixnames.R
\name{unfixnames}
\alias{unfixnames}
\title{prepare column names for presentation}
\usage{
unfixnames(df)
}
\arguments{
\item{df}{dataframe whose colnames you want to standardize}
}
\description{
prepare column names for presentation
}
\examples{
library(smutilities)
df <- iris
names(df)
names(df) <- fixnames(df)
names(df)
names(df) <- unfixnames(df)
names(df)
}

