% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf.R
\name{wf}
\alias{wf}
\title{write dataframe to feather file}
\usage{
wf(df, dfname, directory = here::here("feather"))
}
\arguments{
\item{df}{dataframe you want to save as a feather file}

\item{dfname}{name for the saved file, in quotes, without an extension}

\item{directory}{directory to save file to}
}
\description{
write dataframe to feather file
}
\examples{
library(smutilities)
df <- iris
fd <- here::here("feather")
wf(df, "iris", fd)
}

