% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vartable.R
\name{write_vartable}
\alias{write_vartable}
\title{write_vartable}
\usage{
write_vartable(df, dd = dd, filename)
}
\arguments{
\item{df}{a dataframe with variables as columns and obs as rows}

\item{dd}{the target save directory.  dd by default}

\item{filename}{The filename and file extension, in quotes}
}
\description{
write_vartable
}
\keyword{table}
\keyword{variable}

